+++ homefinder/backend/server.py
@@ -5,20 +5,11 @@ import os
 current_file_directory = os.path.dirname(os.path.realpath(__file__))
 database_dir = os.path.join(current_file_directory,'..','database')
 sys.path.insert(0, database_dir)
-from database import PropertyDatabase
+from database import Database
 
 app = Flask(__name__)
 CORS(app)  # Allow requests from your React app
-property_database = PropertyDatabase()
-row = [1, 'Property name', 'Description', 2, 1, 100, 50000.00, 'Buy', 'Condo', 'garage', 'in_suite', True, False, 0.0000000, 0.0000000, 2000, False, True, True, False]
-property_database.insert_row('property_table', row)
-
-# Insert multiple rows
-rows = [
-    [2, 'Property name 2', 'Description 2', 3, 2, 150, 60000.00, 'Rent', 'Apartment', 'underground', 'shared', False, True, 1.0000000, 1.0000000, 2001, True, False, False, True],
-    [3, 'Property name 3', 'Description 3', 4, 3, 200, 70000.00, 'Buy', 'House', 'covered', 'in_suite', True, False, 2.0000000, 2.0000000, 2002, False, True, True, False]
-]
-property_database.insert_row('property_table', rows)
+property_database = Database()
 @app.route('/send_message', methods=['POST'])
 def send_message():
     message_history = request.get_json()  # Get the message history from the request
