+++ homefinder/database/database.py
@@ -1,22 +1,27 @@
 import os
 import sqlite3
 from typing import List, Optional, Tuple, Any
+import pandas as pd
 
-class PropertyDatabase:
+class Database:
     def __init__(self):
         self.database_name = 'property_db'
-        self._create_property_table()
+        self._create_table()
 
     def get_connection(self):
         conn = sqlite3.connect(self.database_name, uri=True)
         cursor = conn.cursor()
         return conn, cursor
 
-    def _create_property_table(self):
-        query = self.get_property_table_definition()
+    def _create_table(self):
+        query = ''
+        query += self.get_property_table_definition()
+        query += self.get_image_table_definition()
+        query += self.get_agent_table_definition()
+        
         conn, cursor = self.get_connection()
-        cursor.execute('DROP TABLE if exists property_table')
-        cursor.execute(query)
+        cursor.executescript('DROP TABLE if exists property_table;DROP TABLE if exists image_table;DROP TABLE if exists agent_table;'+query)
+        # cursor.execute(query)
         conn.commit()
         conn.close()
 
@@ -30,7 +35,7 @@ class PropertyDatabase:
                     area_size INT,
                     price FLOAT,
                     transaction_type TEXT CHECK( transaction_type IN ('Buy', 'Rent') ),
-                    property_type TEXT CHECK( property_type IN ('Condo', 'Apartment', 'House') ),
+                    property_type TEXT CHECK( property_type IN ('Condo', 'Apartment', 'House', 'Townhouse') ),
                     parking TEXT CHECK( parking IN ('garage', 'underground', 'covered', 'outdoor') ),
                     laundry TEXT CHECK( laundry IN ('in_suite', 'shared') ),
                     furnished BOOL,
@@ -43,6 +48,65 @@ class PropertyDatabase:
                     hardwood_floors BOOL,
                     balcony BOOL
                 );'''
+
+    def get_image_table_definition(self) -> str:
+        return '''CREATE TABLE image_table (
+                image_id int not null,
+                property_id int not null,
+                image_type TEXT CHECK( image_type IN ('indoor', 'outdoor') ),
+                image_url text,
+                PRIMARY KEY (image_id),
+                FOREIGN KEY (property_id) REFERENCES property_table(property_id)
+            );'''
+
+    def get_agent_table_definition(self) -> str:
+        return '''CREATE TABLE agent_table (
+                agend_id int not null,
+                property_id int not null,
+                primary_email text,
+                primary_phone text,
+                PRIMARY KEY (agend_id),
+                FOREIGN KEY (property_id) REFERENCES property_table(property_id)
+            );'''
+
+    def convert_csv_to_sql(self, table_name='property_table', csv_file_name='property_table.csv'):
+        conn, cursor = self.get_connection()
+        sql_script = ""
+        if table_name == 'property_table':
+            df = pd.read_csv(csv_file_name)
+            print (df.columns.tolist())
+            for row in df.itertuples():
+                sql_script = '''
+                    INSERT INTO property_table (property_id, property_name, description, num_bedroom, num_bathroom, 
+                    area_size, price, transaction_type, property_type, parking, laundry, furnished, pet_friendly,
+                    longitude, latitude, build_year, air_conditioning, hardwood_floors, balcony)
+                    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+                    '''
+                    #
+                record = (row.property_id, 
+                    row.property_name,
+                    row.description,
+                    row.num_bedroom,
+                    row.num_bathroom,
+                    row.area_size,
+                    row.price,
+                    row.transaction_type,
+                    row.property_types,
+                    row.parking,
+                    row.laundry,
+                    row.furnished,                    
+                    row.pet_friendly,
+                    row.longitude,
+                    row.latitude,
+                    row.build_year,
+                    row.air_conditioning,
+                    row.hardwood_floors,
+                    row.balcony)
+                cursor.execute(sql_script, record)
+            conn.commit()
+        else:
+            pass
+        return 
 
     def insert_row(self, table_name: str, data: List[List[Any]]):
         conn, cursor = self.get_connection()
@@ -76,3 +140,11 @@ class PropertyDatabase:
         result = cursor.fetchall()
         conn.close()
         return result
+
+
+if __name__ == "__main__":
+    sql_class = Database()
+    sql_class.convert_csv_to_sql( table_name='property_table', csv_file_name='../database/property_table.csv')
+    data = sql_class.fetch_query(query='SELECT * FROM property_table;')
+    print (data)
+    # sql_class.close()
